IOS-6237: Get rid of duplicated `Decimal` extensions in `BSDK` target
