+++ Modules/TangemFoundation/Extensions/Foundation/Decimal+.swift
@@ -25,8 +25,27 @@ public extension Decimal {
         self.init(string: stringValue, locale: .posixEnUS)
     }
 
+    /// return 8 bytes of integer. LittleEndian  format
+    var bytes8LE: [UInt8] {
+        let int64value = (rounded(scale: 0) as NSDecimalNumber).intValue
+        let bytes8 = int64value.bytes8LE
+        return Array(bytes8)
+    }
+
+    var int64Value: Int64 {
+        decimalNumber.int64Value
+    }
+
+    var uint64Value: UInt64 {
+        decimalNumber.uint64Value
+    }
+
     var decimalNumber: NSDecimalNumber {
         self as NSDecimalNumber
+    }
+
+    var roundedDecimalNumber: NSDecimalNumber {
+        rounded(roundingMode: .up) as NSDecimalNumber
     }
 
     var doubleValue: Double {
@@ -46,5 +65,22 @@ public extension Decimal {
         var localCopy = self
         NSDecimalRound(&result, &localCopy, scale, roundingMode)
         return result
+    }
+
+    mutating func round(scale: Int = 0, roundingMode: NSDecimalNumber.RoundingMode = .down) {
+        var localCopy = self
+        NSDecimalRound(&self, &localCopy, scale, roundingMode)
+    }
+
+    func isEqual(to value: Decimal, delta: Decimal) -> Bool {
+        abs(self - value) <= delta
+    }
+
+    func moveRight(decimals: Int) -> Decimal {
+        self * pow(10, decimals)
+    }
+
+    func moveLeft(decimals: Int) -> Decimal {
+        self / pow(10, decimals)
     }
 }
