+++ Tangem/Modules/ExpressView/ExpressViewModel.swift
@@ -11,6 +11,7 @@ import TangemExpress
 import UIKit
 import enum TangemSdk.TangemSdkError
 import struct BlockchainSdk.Fee
+import TangemFoundation
 
 final class ExpressViewModel: ObservableObject {
     // MARK: - ViewState
